scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Kotet <kotet0418@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pdfminer-six = "^20221105"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
